# bff/openapi/widgets.openapi.yaml
openapi: 3.0.0
info:
  title: SDX Reference Implementation BFF - Widgets
  description: >-
    Backend-for-frontend API for the SDX Reference Implementation UI.


    The BFF starts Authorization Code with PKCE login, exchanges the
    authorization code with a confidential client, maintains the user's
    server-side session with an HttpOnly cookie, and proxies Widget calls to the
    SDX-facing provider API.
  version: 0.1.0
  contact:
    name: SDX Reference Implementation Maintainers
  license:
    name: Apache-2.0
servers:
  - url: /
    description: Same-origin BFF API. Operation paths include the /api prefix.
...
paths:
...
  /api/v1/widgets:
    get:
      operationId: listWidgets
      parameters:
        - name: status
          required: false
          in: query
          description: Filter widgets by lifecycle status.
          schema:
            example: active
            type: string
            enum:
              - active
              - inactive
              - archived
        - name: name
          required: false
          in: query
          description: Case-insensitive partial name match.
          schema:
            minLength: 1
            maxLength: 200
            example: intake
            type: string
        - name: limit
          required: false
          in: query
          description: Maximum number of widgets to return.
          schema:
            minimum: 1
            maximum: 100
            default: 25
            example: 25
            type: integer
        - name: cursor
          required: false
          in: query
          description: Opaque pagination cursor from the previous response.
          schema:
            example: eyJvZmZzZXQiOjI1fQ==
            type: string
        - name: sortBy
          required: false
          in: query
          description: Field used to sort results.
          schema:
            example: createdAt
            type: string
            enum:
              - createdAt
              - updatedAt
              - name
              - status
        - name: sortOrder
          required: false
          in: query
          description: Sort direction for the selected sort field.
          schema:
            example: desc
            type: string
            enum:
              - asc
              - desc
      responses:
        '200':
          description: The list of Widgets owned by the signed-in user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WidgetListResponse'
        '401':
          description: Login is required.
      security:
        - bff_session: []
      summary: List Widgets for the signed-in user through the BFF.
      tags:
        - BFF Widgets
    post:
      operationId: createWidget
      parameters:
        - name: Idempotency-Key
          in: header
          description: Optional client-generated key forwarded to the provider SDX API.
          required: false
          schema:
            type: string
            minLength: 8
            maxLength: 255
            example: req-12345678
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWidgetRequest'
      responses:
        '201':
          description: The created Widget.
          headers:
            ETag:
              description: Entity tag returned by the provider SDX API.
              schema:
                type: string
                example: '"u6I3AI8rSnvR3uOSYVQbPiZF7cP8fIQ77U1zba2tI8A"'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Widget'
        '401':
          description: Login is required.
      security:
        - bff_session: []
      summary: Create a Widget for the signed-in user through the BFF.
      tags:
        - BFF Widgets
  '/api/v1/widgets/{widgetId}':
    get:
      operationId: getWidget
      parameters:
        - name: widgetId
          required: true
          in: path
          description: The UUID of the Widget.
          schema:
            type: string
            format: uuid
            example: 4f3066e8-5a59-4fc5-8e7b-fcd7f4d01c4f
      responses:
        '200':
          description: The requested Widget.
          headers:
            ETag:
              description: Entity tag returned by the provider SDX API.
              schema:
                type: string
                example: '"u6I3AI8rSnvR3uOSYVQbPiZF7cP8fIQ77U1zba2tI8A"'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Widget'
        '401':
          description: Login is required.
      security:
        - bff_session: []
      summary: Get a Widget for the signed-in user through the BFF.
      tags:
        - BFF Widgets
    put:
      operationId: replaceWidget
      parameters:
        - name: widgetId
          required: true
          in: path
          description: The UUID of the Widget.
          schema:
            type: string
            format: uuid
            example: 4f3066e8-5a59-4fc5-8e7b-fcd7f4d01c4f
        - name: If-Match
          in: header
          description: >-
            Optional entity tag from a previous GET response. Forwarded to the
            provider SDX API.
          required: false
          schema:
            type: string
            example: '"u6I3AI8rSnvR3uOSYVQbPiZF7cP8fIQ77U1zba2tI8A"'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWidgetRequest'
      responses:
        '200':
          description: The replaced Widget.
          headers:
            ETag:
              description: Entity tag returned by the provider SDX API.
              schema:
                type: string
                example: '"u6I3AI8rSnvR3uOSYVQbPiZF7cP8fIQ77U1zba2tI8A"'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Widget'
        '401':
          description: Login is required.
      security:
        - bff_session: []
      summary: Replace a Widget for the signed-in user through the BFF.
      tags:
        - BFF Widgets
    patch:
      operationId: updateWidget
      parameters:
        - name: widgetId
          required: true
          in: path
          description: The UUID of the Widget.
          schema:
            type: string
            format: uuid
            example: 4f3066e8-5a59-4fc5-8e7b-fcd7f4d01c4f
        - name: If-Match
          in: header
          description: >-
            Optional entity tag from a previous GET response. Forwarded to the
            provider SDX API.
          required: false
          schema:
            type: string
            example: '"u6I3AI8rSnvR3uOSYVQbPiZF7cP8fIQ77U1zba2tI8A"'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchWidgetRequest'
      responses:
        '200':
          description: The partially updated Widget.
          headers:
            ETag:
              description: Entity tag returned by the provider SDX API.
              schema:
                type: string
                example: '"u6I3AI8rSnvR3uOSYVQbPiZF7cP8fIQ77U1zba2tI8A"'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Widget'
        '401':
          description: Login is required.
      security:
        - bff_session: []
      summary: Partially update a Widget for the signed-in user through the BFF.
      tags:
        - BFF Widgets
    delete:
      operationId: deleteWidget
      parameters:
        - name: widgetId
          required: true
          in: path
          description: The UUID of the Widget.
          schema:
            type: string
            format: uuid
            example: 4f3066e8-5a59-4fc5-8e7b-fcd7f4d01c4f
        - name: If-Match
          in: header
          description: >-
            Optional entity tag from a previous GET response. Forwarded to the
            provider SDX API.
          required: false
          schema:
            type: string
            example: '"u6I3AI8rSnvR3uOSYVQbPiZF7cP8fIQ77U1zba2tI8A"'
      responses:
        '204':
          description: The Widget was deleted.
        '401':
          description: Login is required.
      security:
        - bff_session: []
      summary: Delete a Widget for the signed-in user through the BFF.
      tags:
        - BFF Widgets
components:
  securitySchemes:
    bff_session:
      type: apiKey
      in: cookie
      name: bff_session
      description: >-
        HttpOnly BFF session cookie created by /api/auth/login. Swagger users
        should log in through the application first; the browser will then send
        this cookie automatically with same-origin Swagger requests.
  schemas:
    WidgetStatus:
      type: string
      enum:
        - active
        - inactive
        - archived
    WidgetSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 4f3066e8-5a59-4fc5-8e7b-fcd7f4d01c4f
          readOnly: true
        subject:
          type: string
          description: Owner subject from the provider SDX API.
          example: user-123
          readOnly: true
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Intake form
        status:
          example: active
          allOf:
            - $ref: '#/components/schemas/WidgetStatus'
        updatedAt:
          format: date-time
          type: string
          example: '2026-05-13T18:00:00Z'
          readOnly: true
      description: Summary representation used in widget list responses.
      required:
        - id
        - subject
        - name
        - status
        - updatedAt
    WidgetListResponse:
      type: object
      properties:
        items:
          example:
            - id: 4f3066e8-5a59-4fc5-8e7b-fcd7f4d01c4f
              subject: user-123
              name: Intake form
              status: active
              updatedAt: '2026-05-13T18:00:00Z'
          type: array
          items:
            $ref: '#/components/schemas/WidgetSummary'
        nextCursor:
          type: string
          nullable: true
          description: >-
            Cursor to retrieve the next page, or null when there are no more
            results.
          example: eyJvZmZzZXQiOjI1fQ==
      description: >-
        A paginated list of widget summaries with an optional cursor for the
        next page.
      required:
        - items
    CreateWidgetRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Intake form
        description:
          type: string
          nullable: true
          maxLength: 1000
          example: Widget used for intake workflow testing.
        status:
          default: active
          example: active
          allOf:
            - $ref: '#/components/schemas/WidgetStatus'
        additionalData:
          type: object
          additionalProperties: true
          example:
            source: local-dev
      description: >-
        Request body for creating a widget. Subject is never accepted here. If
        status is omitted, the service creates the widget with status `active`.
        If additionalData is omitted, the service stores an empty additional
        data object.
      required:
        - name
    Widget:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 4f3066e8-5a59-4fc5-8e7b-fcd7f4d01c4f
          readOnly: true
        subject:
          type: string
          description: Owner subject from the provider SDX API.
          example: user-123
          readOnly: true
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Intake form
        description:
          type: string
          nullable: true
          maxLength: 1000
          example: Widget used for intake workflow testing.
        status:
          example: active
          allOf:
            - $ref: '#/components/schemas/WidgetStatus'
        additionalData:
          type: object
          additionalProperties: true
          example:
            source: local-dev
        createdAt:
          format: date-time
          type: string
          example: '2026-05-13T18:00:00Z'
          readOnly: true
        updatedAt:
          format: date-time
          type: string
          example: '2026-05-13T18:00:00Z'
          readOnly: true
      description: Widget resource owned by a subject.
      required:
        - id
        - subject
        - name
        - description
        - status
        - additionalData
        - createdAt
        - updatedAt
    UpdateWidgetRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Intake form
        description:
          type: string
          nullable: true
          maxLength: 1000
          example: Widget used for intake workflow testing.
        status:
          default: active
          example: active
          allOf:
            - $ref: '#/components/schemas/WidgetStatus'
        additionalData:
          type: object
          additionalProperties: true
          example:
            source: local-dev
      description: >-
        Request body for replacing a widget. Subject is never accepted here.
        This is a full replacement. Omitted optional fields are reset to their
        default values: description is set to null, status is set to active, and
        additionalData is set to an empty object.
      required:
        - name
    PatchWidgetRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Intake form
        description:
          type: string
          nullable: true
          maxLength: 1000
          example: Widget used for intake workflow testing.
        status:
          default: active
          example: active
          allOf:
            - $ref: '#/components/schemas/WidgetStatus'
        additionalData:
          type: object
          additionalProperties: true
          example:
            source: local-dev
      description: >-
        Request body for partially updating a widget. Subject is never accepted
        here. Omitted fields preserve their existing values.